requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg[binary]>=3.3.3
pydantic-settings==2.5.2
email-validator>=2.2.0
python-dotenv==1.0.1
supabase==2.9.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
weasyprint==62.3
twilio==9.3.5
apscheduler==3.10.4
python-dateutil==2.9.0
pytz==2024.2
httpx==0.27.2
jinja2>=3.1.0
